Generate Passengers

+++ Assets/Scripts/Scriptable/FloorData.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New Floor")]
 public class Floor : ScriptableObject
 {
-    public float FloorNumber;
+    public int FloorNumber;
     public Sprite ButtonSprite;
     public string Name;
 }
